pyproject.toml
[project]
name = "langchain-rag-examples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "langchain==0.3.26",
    "langchain-ollama",
    "langchain-chroma",
    "pandas",
    "python-dotenv",
    "anthropic"
]

[project.scripts]
rag-examples = "rag_examples.main:main"
